+++ cmd/server/process_unix.go
@@ -1,0 +1,19 @@
+//go:build unix
+
+package server
+
+import (
+	"errors"
+	"os"
+	"syscall"
+)
+
+// isProcessRunning checks if a process with the given PID is running.
+func isProcessRunning(pid int) bool {
+	process, err := os.FindProcess(pid)
+	if err != nil {
+		return false
+	}
+	err = process.Signal(syscall.Signal(0))
+	return err == nil || errors.Is(err, syscall.EPERM)
+}